Make crud list actions overridable/extendable
Crud views automatically get delete and edit actions from models, which
are shown when listing the objects. This commit makes it possible to
override these actions, or add new ones to extend the existing
functionality.

This is done by adding `crud_override_actions` and `crud_extra_actions`
attributes to the Meta class of the models.