Fix permission code and name generation
Permissions of tasks had codes in the form of `workflow.task`. This was
instead changed to `workflow.lane.task`.

Lanes used a 'permissions' extension on the lane as their permission
codes. This was scrapped in favour of the lane ids. Lane permissions are
now generated in the form of `workflow.lane`.

Finally, changes have been made to the way permission names are
generated. While all permissions simply used to utilize ids to create
their names, now the actual 'name' field of the diagram is used.