Add status of the role on the permission tree view
The status of the role has been added on the permission tree view, so
that the elements of the permission tree will have a "checked" attribute
that denotes whether the selected role has that permission or not.

Also, the add-remove permission step of the workflow has been added as
well. To perform this addition, a new method has been defined on the
role object to be able to remove permissions from a role.