Fix i18n import, add pluralization/marking
Internationalization support previously relied on the '_' method being
installed into the builtins namespace, which ZEngine did when
running. However, this meant that any code that used the i18n support
could not be used in places that did not run a ZEngine instance, such as
the management commands. The functions now need to be explicitly
imported, the commands can now be used everywhere.

Additionally, support for pluralization has been added. This makes it
easier to translate messages that have both singular and plural forms,
as well as making it possible to translate to the languages that support
multiple plural forms, or use different pluralization rules.