Add language negotiation over multiple prefs
The field 'lang_code' was replaced with 'lang_codes'. The field now a
list of user preferences. When negotiating the language preferences with
the available languages, a suitable one is picked among the
preferences. The rationale of this change is that the browsers supply
the language preferences as a list, such as `['tr_TR', 'tr', 'en_US',
'en']`.