Fix locale to keep it on session/user
The locale preference of the user is now kept on the session, as well as
being stored on the User model.

The preferences of language, date formats and number formats have been
split. This makes it possible for the user to use one language while
having the dates formatted as a different locale, for example using
Turkish language while having the dates formatted in American English
style.

A primitive workflow has been added for changing the language. This is
mainly for testing purposes only, and will later be replaced with a more
robust solution.

rref #5409