Fix translations for unicode messages on python2
In Python 2, `gettext` only accepts ASCII characters in message ids. If
the message ids contain unicode characters, it is unable to match these
messages with the translations. However, if the messages are marked as
unicode, then the encoding of the translation files and the messages
match, and the translations work.

To fix this problem, the translation functions have been switched with
their unicode variants, and the documentation has been updated to
clarify the need for the messages to be unicode.

rref #5409