Add domain support to i18n & ZEngine translations
Added domain support to the internationalization system. This is used to
split the translations into different files, which makes it possible to
use different translations for the same word in different parts of the
application. For example the word 'date' may be used in both 'romantic
date' meaning, and 'calendar date' meaning within the same application,
in which case the correct translations to different languages could only
be done by splitting both usages to seperate domains.

Another feature of the domain support is that the implementation in
ZEngine makes it possible to use different default languages for the
messages in different parts of the application. For example, some of the
application may have English messages while the rest uses Turkish
messages, as long as both parts use different domains this will work
correctly.

Finally, the domain support was used to add translations for ZEngine
itself. Applications supporting languages other than English can add
their own translations of messages used by ZEngine by placing the
translations under 'zengine' domain.