Fix loading when no translation is created yet
While the management commands are required to create the translation
files, if there are no translation files already in place, importing
`zengine.lib.translation` would crash. Translation module is fixed to
instead log an error if the language doesn't exist, making it possible
to use the management commands to create the language files.